# Setup for benchmarks, performance optimizations