--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -14,13 +14,11 @@ on:
     branches: [ master ]
 
 jobs:
-  test:
-
+  lint:
     runs-on: ubuntu-latest
     strategy:
       matrix:
         ruby: [ '2.6', '2.7', '3.0' ]
-
     steps:
     - uses: actions/checkout@v2
     - name: Set up Ruby