--- a/Gemfile
+++ b/Gemfile
@@ -12,3 +12,5 @@ gem 'redcarpet'
 gem 'yard-activesupport-concern'
 
 gem 'solargraph'
+
+gem 'benchmark-ips'