--- a/Gemfile
+++ b/Gemfile
@@ -14,3 +14,5 @@ gem 'yard-activesupport-concern'
 gem 'solargraph'
 
 gem 'benchmark-ips'
+
+gem 'rspec-its'