--- a/bench/benchmark.rb
+++ b/bench/benchmark.rb
@@ -1,0 +1,33 @@
+require 'bundler/setup'
+
+require 'sober_swag'
+
+require 'yaml'
+
+##
+# Quick and dirty way to benchmark things.
+class Bench
+  class << self
+    def report(name, &block)
+      puts name
+
+      data[name] ||= Benchmark.ips(&block).data
+    end
+
+    def data
+      @data ||= {}
+    end
+
+    def write!(filename)
+      File.open(filename, 'w') do |f|
+        f << YAML.dump(data)
+      end
+    end
+  end
+end
+
+Dir['bench/benchmarks/**/*.rb'].sort.each do |file|
+  require file
+end
+
+Bench.write!('benchmark_results.yaml')