--- a/lib/sober_swag/serializer.rb
+++ b/lib/sober_swag/serializer.rb
@@ -10,6 +10,7 @@ module SoberSwag
     autoload(:Mapped, 'sober_swag/serializer/mapped')
     autoload(:Optional, 'sober_swag/serializer/optional')
     autoload(:FieldList, 'sober_swag/serializer/field_list')
+    autoload(:Hash, 'sober_swag/serializer/hash')
     autoload(:Meta, 'sober_swag/serializer/meta')
 
     class << self