This PR:

1. Adds a way to write benchmarks for SoberSwag, and runs those benchmarks on CI.
2. Changes the field serializer to use mutation. On my machine, this resulted in the following speedups:
  ```
  BASIC SERIALIZATION USING OUTPUT OBJECTS:

  Serializer of Output Object After:   927283.5 i/s
   Output Object After:   856398.1 i/s - 1.08x  (± 0.00) slower
  Serializer of Output Object Before:   825511.4 i/s - 1.12x  (± 0.00) slower
  Output Object Before:   686776.6 i/s - 1.35x  (± 0.00) slower
  ```
3. Changes the field serializer to use a new, hash-dispatch-based serializer for view selection. This resulted in the following speedups:

```
Comparison:
  With no view after:   984076.1 i/s
   With a view after:   423094.0 i/s - 2.33x  (± 0.00) slower
  With a view before:   195679.2 i/s - 5.03x  (± 0.00) slower
 With no view before:   153382.0 i/s - 6.42x  (± 0.00) slower
```